+++ html/js/portal-ng.js
@@ -30,7 +30,7 @@ THE SOFTWARE.
  former: Portal-NG (PoNG) https://mh-svr.de/mw/index.php/PoNG
 */
 var labeldefs = new Array();
-var PONGVER = '2.17.2';
+var PONGVER = '2.18.0';
 labeldefs['PONGVER'] = PONGVER;
 
 var moduleMap = {};